day 9
Simple calculation first time round.

for the second question we paint in the peaks first,
then we iterate over the 2d array painting any
empty square.

Whenever an empty square is painted, it immediately
tries to paint it's neighbours. The painting continues
recursively until a non-empty square is found.

at the end just count the occurrences of each painter
and the problem solves itself.