Optimise day 12
I was unhappy with the performance, so i took the time to
remove things i (now) know I didn't need. I wanted to track
all discovered routes in q1 as i guessed that the part 2
would require knowledge of each viable route.

As that's not necessary, remove the current route tracking
and also remove the hashtable of discovered routes.

Finally, swizzle the code around a bit to simplify how
'visit twice' is handled.

Reduces runtime to a fraction of a second.

If this was production code I'd implement some form of
Visitor pattern (pass a lambda? or a structured object
so I can cache whether or not a small room has been
entered twice so i can reduce an O(N) loop into a simple
boolean access?).

Maybe I'll do that later :p