day7.
I started by counting how many unique ranks there were and then from
that working out the strength of the hand. if there were 3 unique ranks
then this was a three-of-a-kind or two pairs. Adding in wildcards
made this an awful lot of if/else.

Rewrote it to be a frequency counter instead, and then wildcards
could be easily pre-processed by 'converting' them into a card of
the appropriate type. For example, if there were two pairs, and
there were 2 wildcards, i knew that wildcards should be removed
from 'bucket 2' first, and then i should add them to the highest
non-empty bucket, which appened to be bucket 2. "Adding" is done
by moving one item in bucket 2 into bucket 4. Problem solved!