d8p2 - done
I read and reread it trying to understand if the rise/run should be
the distance between the two towers, or if I should be trying to make
interior antinodes.

Looks like there are no interior antinodes in part 2? If two towers
were 4 apart in both x and y, then from my reading of the question there
should be an antinode at every single point in the diagonal between
them, and also on the exterior at every node.

Maybe this just doesn't exist in the data and i got lucky?