d20p1
As i was implementing it i realised that an easier approach would be to
just calculate the full path once, and then just see if point N and M
are 2 apart, and if they are, the distance saved is just subtracting
their position on the path.

There is only one path, and the only two actions are:
1) Follow the path
2) Skip 2 spots.

After reading part 2 this seems like a better approach. Now i need to
either
1) Follow the path
2) Skip up to 20 spots (or whatever it is)