d20p2
I spent a while debugging and then realised i was treating the cheat as
teleport, not a 'drive through the wall' type thing.

Tweaked the calculation to take into account the manhatten distance
between the two points when calculating the saving. If we cheat our
way to a node 100 in front of us, but it takes 10 moves to get there,
the saving is only 90 moves.