d16p2 complete
As per prior commit - the path we follow is now a tuple of
[position,direction] rather than just [position].

This means multiple paths can converge to the same trail after the reach
the same node *and* are facing in the same direction.

FOr example, if you enter a node from the west and south, but the
shortest path is to the east, the variant which enters from the west can
'just' walk straight. The variant that enters from the south must rotate
once (adding 1000) to it's score, and then steps to the east.

If the southern variant enters with a score of 1001, it'll leave with
2002. If the western variant enters with score 2001, it'll leave with
score 2002. From that point on the paths are identical. HOwever,
disambiguating this is important when we're backtracking to find *all*
the good paths.